Bug and selling Chips
When you try to sell, no matter if you have chips with you or not, system always sell chips.

Ex. If i have no chips on me, when i hit sell chips its not validating the item so, the system process and say that you sold your chips for $0, of course because i have no chips on me, so i fixed, maybe can be better but atleast for me works.